Revisions based on code review. Back end applied
