+++ packages/commons/src/MAR/CommunityRiskAssessment/CapacityAndVulnerability.js
@@ -1,0 +1,21 @@
+
+const DeleteCapacityAndVulnerability = () => {
+    fetch(`${AppConfig.HOSTNAME}/v2/mar_risk_assessment/cav/remove`, {
+        method: 'DELETE',
+        headers: {
+            Accept: 'application/json',
+            'Content-Type': 'application/json',
+        },
+        body: JSON.stringify({
+            "cav_id": cavID,
+            "site_id": cookies.credentials.site_id
+        }),
+        }).then((response) => response.json())
+        .then((responseJson) => {
+            return responseJson;
+        })
+        .catch((error) => {
+            console.error("Problem in Delete CAV")
+        }
+    );
+};
