+++ packages/server/connections/__init__.py
@@ -90,13 +90,8 @@ def create_app():
     from src.api.v2.users.user_management import USER_MANAGEMENT_BLUEPRINT
     app.register_blueprint(USER_MANAGEMENT_BLUEPRINT, url_prefix="/v2")
 
-    from src.api.v2.marirong.risk_assessment.cra import COMMUNITY_RISK_ASSESSMENT_BLUEPRINT
-    app.register_blueprint(COMMUNITY_RISK_ASSESSMENT_BLUEPRINT, url_prefix="/v2/mar_risk_assessment/cra")
+    from src.api.v2.marirong.community_risk_assessment import COMMUNITY_RISK_ASSESSMENT_BLUEPRINT
+    app.register_blueprint(COMMUNITY_RISK_ASSESSMENT_BLUEPRINT, url_prefix="/v2")
 
-    from src.api.v2.marirong.risk_assessment.cav import CAPACITY_AND_VULNERABILITY_BLUEPRINT
-    app.register_blueprint(CAPACITY_AND_VULNERABILITY_BLUEPRINT, url_prefix="/v2/mar_risk_assessment/cav")
-
-    from src.api.v2.marirong.risk_assessment.hazard_maps import HAZARD_MAPS_BLUEPRINT
-    app.register_blueprint(HAZARD_MAPS_BLUEPRINT, url_prefix="/v2/mar_risk_assessment/hazard_maps")
     
     return app
