+++ packages/web/src/components/marirong/features/CapacityAndVulnerability.js
@@ -27,6 +27,8 @@ import Snackbar from '@material-ui/core/Snackbar';
 import MuiAlert from '@material-ui/lab/Alert';
 import { useCookies } from 'react-cookie';
 
+import { MarCommunityRiskAssessment } from '@dynaslope/commons';
+
 function Alert(props) {
 	return <MuiAlert elevation={6} variant="filled" {...props} />;
 }
@@ -133,6 +135,16 @@ function CapacityAndVulerability() {
     }
 
     const confirmDelete = () => {
+        const input = {
+            "cav_id": cavID,
+            "site_id": cookies.credentials.site_id
+        };
+
+        
+        setTimeout(async () => {
+            const response = await 
+        }, 400);
+
         fetch(`${AppConfig.HOSTNAME}/v2/mar_risk_assessment/cav/remove`, {
             method: 'DELETE',
             headers: {
@@ -159,11 +171,6 @@ function CapacityAndVulerability() {
                     setNotifText("Failed to delete capacity and vunerability data. Please contact the developers or file a bug report");
                 }
             })
-            .catch((error) => {
-                setOpenNotif(true);
-                setNotifStatus("error");
-                setNotifText(`ERROR: ${error}`);
-            }
         );
     }
